+++ api/dc/base/serializers.py
@@ -203,6 +203,7 @@ class DcSettingsSerializer(s.InstanceSerializer):
         'SMS_SMSAPI_PASSWORD',
     })
     _null_fields_ = frozenset({
+        'VMS_VM_DEFINE_LIMIT',
         'VMS_VM_SNAPSHOT_DEFINE_LIMIT',
         'VMS_VM_SNAPSHOT_LIMIT_AUTO',
         'VMS_VM_SNAPSHOT_LIMIT_MANUAL',
@@ -271,6 +272,10 @@ class DcSettingsSerializer(s.InstanceSerializer):
     VMS_ZONE_ENABLED = s.BooleanField(label='VMS_ZONE_ENABLED',  # Module
                                       help_text=_('Whether to enable support for SunOS and Linux zones in '
                                                   'this virtual datacenter.'))
+
+    VMS_VM_DEFINE_LIMIT = s.IntegerField(label='VMS_VM_DEFINE_LIMIT', required=False,
+                                         help_text=_('Maximum number of virtual servers that can be defined in '
+                                                     'this virtual datacenter.'))
     VMS_VM_STOP_TIMEOUT_DEFAULT = s.IntegerField(label='VMS_VM_STOP_TIMEOUT_DEFAULT',
                                                  help_text='Default time period (in seconds) for a graceful VM stop or '
                                                            'reboot, after which a force stop/reboot is send to the VM '